`Autograd` dispatcher for `segment_matmul` (#122)
* update

* changelog

* update
