+++ pyg_lib/csrc/ops/matmul.h
@@ -7,19 +7,16 @@
 namespace pyg {
 namespace ops {
 
-using torch::autograd::Variable;
-using torch::autograd::variable_list;
-
 // Performs matrix multiplication across list of elements.
 // TODO (matthias) Support `out` argument.
-PYG_API std::vector<at::Tensor> grouped_matmul(const variable_list input,
-                                               const variable_list other);
+PYG_API std::vector<at::Tensor> grouped_matmul(const at::TensorList input,
+                                               const at::TensorList other);
 
 // Performs matrix multiplication according to segments.
 // TODO (matthias) Support `out` argument.
-PYG_API at::Tensor segment_matmul(const Variable input,
+PYG_API at::Tensor segment_matmul(const at::Tensor& input,
                                   const at::Tensor& ptr,
-                                  const Variable other);
+                                  const at::Tensor& other);
 
 }  // namespace ops
 }  // namespace pyg
